add searching as a function in projectapi
